requirements.txt
pytest==7.3.2
selene>=2.0.b4
selenium~=4.13.0
pip~=23.1.2
attrs~=23.1.0
build~=0.10.0
keyring~=24.2.0
setuptools~=68.2.2
packaging~=23.1
zipp~=3.17.0
future~=0.18.3
python-dotenv~=1.0.0
allure-pytest